fix(gsd): route slice summary projection write through single-writer layer
